Move new message handler into folder
