+++ editor/src/messages/input_mapper/key_mapping/mod.rs
@@ -1,0 +1,7 @@
+mod key_mapping_message;
+mod key_mapping_message_handler;
+
+#[doc(inline)]
+pub use key_mapping_message::{KeyMappingMessage, KeyMappingMessageDiscriminant, MappingVariant, MappingVariantDiscriminant};
+#[doc(inline)]
+pub use key_mapping_message_handler::KeyMappingMessageHandler;
