+++ editor/src/messages/input_mapper/mod.rs
@@ -1,16 +1,11 @@
 mod input_mapper_message;
 mod input_mapper_message_handler;
-mod key_mapping_message;
-mod key_mapping_message_handler;
 
 pub mod default_mapping;
+pub mod key_mapping;
 pub mod utility_types;
 
 #[doc(inline)]
 pub use input_mapper_message::{InputMapperMessage, InputMapperMessageDiscriminant};
 #[doc(inline)]
 pub use input_mapper_message_handler::InputMapperMessageHandler;
-#[doc(inline)]
-pub use key_mapping_message::{KeyMappingMessage, KeyMappingMessageDiscriminant, MappingVariant, MappingVariantDiscriminant};
-#[doc(inline)]
-pub use key_mapping_message_handler::KeyMappingMessageHandler;
