+++ editor/src/messages/input_mapper/utility_types/misc.rs
@@ -1,6 +1,6 @@
 use super::input_keyboard::{all_required_modifiers_pressed, KeysGroup, LayoutKeysGroup};
+use crate::messages::input_mapper::key_mapping::MappingVariant;
 use crate::messages::input_mapper::utility_types::input_keyboard::{KeyStates, NUMBER_OF_KEYS};
-use crate::messages::input_mapper::MappingVariant;
 use crate::messages::prelude::*;
 
 use serde::{Deserialize, Serialize};
